refactor and documentation update
